requirements.txt
instagrapi
requests
Pillow
praw
openai
